xray2model/requirements.txt
numpy>=1.20.0
scipy>=1.7.0
torch>=1.9.0
torchvision>=0.10.0
Flask>=2.0.0
flask-cors>=3.0.10
pillow>=8.2.0
scikit-image>=0.18.0
matplotlib>=3.4.0
pandas>=1.3.0
tensorflow>=2.6.0
SimpleITK>=2.1.0
nibabel>=3.2.0
vtk>=9.0.0
monai>=1.0.0
trimesh>=3.0.0 
fast_simplification>=0.1 # Add fast_simplification explicitly
torchdr>=0.1.4 # Add torchdr for DRR simulation
